[TEMP] Start on supporting '--dispatch-namespace' in local dev
This registers this worker as a namespaced worker in the dev registry.
Kind of a bit of a hack, as it's just registered with the name
'<namespace>:<worker>'.

Up next is trying to figure out how to build a dispatch namespace plugin
for a namespace binding which just proxies to the registered workers